+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ### 2021-07-02/kw
 ### Added
+- new routes added to manage apikeys (activate/deactivate apikeys)
+
+### 2021-07-02/kw
+### Added
 - New Pivot Table api_key_institution added . ManyToMany between api_key and institution
 - Middleware AuthorizeApiKey.php to authorize request to add/update external offers added
 - Models Institution & Apikey changed to have manyToMany relationship via pivot table